requirements.txt
Pillow==10.2.0
PySide6==6.6.1
pyinstaller==6.3.0 